Fail to redesign layers.py and net.py (not pretty)
I give up on redesigning the layer classes. They just make the problem
more confusing. Layer classes are included in this commit, but the next
they will be gone.

Usually classes simplify a program by acting as ONE distinct idea.
That is impossible for a layer class. For a programmer to use it she
must know the order to call methods for propagation and
backpropagation, which takes thinking on the level of the neurons,
defeating the entire purpose.